Refactor GraphQL resolvers
