+++ backend/resolvers/resolvers.root.js
@@ -1,0 +1,16 @@
+const UsersQuery = require('./users/users.query');
+const UsersResolver = require('./users/users.resolver');
+const TestQuery = require('./test.query');
+
+const resolvers = {
+  Query: {
+    ...UsersQuery,
+    ...TestQuery,
+  },
+
+  Mutation: {
+    ...UsersResolver,
+  }
+};
+
+module.exports = resolvers;
