+++ backend/resolvers/test.query.js
@@ -1,0 +1,10 @@
+const TestQuery = {
+  test() {
+    return 'cool beans!'
+  },
+  reverse(_, args) {
+    return args.word.split("").reverse().join("")
+  }
+}
+
+module.exports = TestQuery;
