+++ backend/resolvers/users/users.data.js
@@ -1,0 +1,14 @@
+const users = [
+  {
+    name: "test1",
+    email: "email1",
+    id: '1',
+  },
+  {
+    name: "test2",
+    email: "email2",
+    id: '2',
+  },
+];
+
+module.exports = users;
