+++ backend/resolvers/users/users.query.js
@@ -1,0 +1,7 @@
+const users = require('./users.data');
+
+const UsersQuery = {
+  users: () => users,
+}
+
+module.exports = UsersQuery
